Move flattenObject to separate util module
