+++ webapp/src/app/api/translations/[lang]/route.ts
@@ -1,4 +1,5 @@
 import { getLanguage } from '@/app/api/languages';
+import flattenObject from '@/utils/flattenObject';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function GET(
@@ -13,28 +14,4 @@ export async function GET(
     lang,
     translations: flattenLangObj,
   });
-}
-
-function flattenObject(
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  obj: Record<string, any>,
-  parentKey: string = '',
-): Record<string, string> {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const result: Record<string, any> = {};
-
-  for (const key in obj) {
-    // eslint-disable-next-line no-prototype-builtins
-    if (obj.hasOwnProperty(key)) {
-      const newKey = parentKey ? `${parentKey}.${key}` : key;
-
-      if (typeof obj[key] === 'object' && !Array.isArray(obj[key])) {
-        Object.assign(result, flattenObject(obj[key], newKey));
-      } else {
-        result[newKey] = obj[key];
-      }
-    }
-  }
-
-  return result;
 }
